@scylladb-promoter 2023-11-07
#### :red_circle: CI State: FAILURE

:x: - Build
#### Build Details:
- Build URL: https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/4559/
- Duration: 13 min
- Builder: spider4.cloudius-systems.com