@mykaul 2023-11-07
> #### 🔴 CI State: FAILURE
> ❌ - Build
> 
> #### Build Details:
> * Build URL: https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/4559/
> * Duration: 13 min
> * Builder: spider4.cloudius-systems.com

```
ld.lld: error: undefined symbol: flush_rows(seastar::lw_shared_ptr<schema const>, std::__cxx11::list<repair_row, std::allocator<repair_row>>&, seastar::lw_shared_ptr<repair_writer>&)
 >>> referenced by repair_test.cc
 >>>               build/dev/test/boost/repair_test.o:(seastar::noncopyable_function<void ()>::direct_vtable_for<seastar::futurize<std::invoke_result<flush_repair_rows_on_wire_to_sstable::run_test_case() const::$_0>::type>::type seastar::async<flush_repair_rows_on_wire_to_sstable::run_test_case() const::$_0>(seastar::thread_attributes, flush_repair_rows_on_wire_to_sstable::run_test_case() const::$_0&&)::'lambda'()>::call(seastar::noncopyable_function<void ()> const*))
 clang-16: error: linker command failed with exit code 1 (use -v to see invocation)
```