@tzach 2023-11-07
> Users only need to send the repair rest api to one of the nodes in the cluster. It can be any of the nodes in the cluster.

What happens if the user, for example, current versions of Scylla Manager, sends multiple repair requests?