@tzach 2023-11-07
Why 100MiB per node? The test above is much smaller. Did you test with 100MiB to make sure it work as expected?