@tzach 2023-11-07
How should the user (or Scylla Manager) know if he should use the small table option?
The best will be to have this decision to Scylla or to document an API to estimate table size.
Another alternative is to assume all system tables are small, but then we will miss a small table created by the user.