@asias 2023-11-07
> Why 100MiB per node? The test above is much smaller. Did you test with 100MiB to make sure it work as expected?

100 MiB is just a number. Table of any size should work. But it is less efficient than the regular repair without this optimization if the amount of data is big. Yes, I tested with GBs of data.