@asias 2023-11-07
> Do we need tests for this feature?

Yes, I have local dtest which is not upstreamed.