@asias 2023-11-07
> How should the user (or Scylla Manager) know if he should use the small table option? The best will be to have this decision to Scylla or to document an API to estimate table size. Another alternative is to assume all system tables are small, but then we will miss a small table created by the user.

There is no simple formula to give a definite answer when it is more efficient to turn on this flag for different cluster sizes and table sizes.  But for sure, it is more efficient for tables with tiny amount of data like the ones with system tables. 

We can turn it on for a given threshold. It is not a big deal if we miss some user created small tables.  We can also run with/without this optimization in different runs in Scylla manager to get a hint if scylla manager should turn it on.