@tzach 2023-11-07
> We can turn it on for a given threshold. 

A threshold on what metric?

This improvement is a huge win, but without automation in either Scylla or Scylla Manager, it would be less valuable. 
As a start, Manager can assume system and audit keyspaces are small.