@mykaul 2023-11-07
> As a start, Manager can assume system and audit keyspaces are small.

I think all system tables are small.