@scylladb-promoter 2023-11-07
#### :green_circle: CI State: SUCCESS

:white_check_mark: - Build
:white_check_mark: - Sanity Tests
:white_check_mark: - Unit Tests
#### Build Details:
- Build URL: https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/4560/
- Duration: 2 hr 11 min
- Builder: spider4.cloudius-systems.com