@avikivity 2023-11-07
> The time to finish repair difference = 1.5648403 seconds / 0.001459798 seconds = 1072X

But it's just 1.5 seconds faster.

If it's 20 minutes -> 1.2 seconds, it's a big advantage.

But 1.5 seconds -> 1.5ms isn't user visible.

So we need better justification.