@mykaul 2023-11-07
I was impressed by the reduction in RPC calls.
rpc_call_nr=115289 (vs 80 after)