@asias 2023-11-08
> > The time to finish repair difference = 1.5648403 seconds / 0.001459798 seconds = 1072X
> 
> But it's just 1.5 seconds faster.
> 
> If it's 20 minutes -> 1.2 seconds, it's a big advantage.
> 
> But 1.5 seconds -> 1.5ms isn't user visible.
> 
> So we need better justification.

The pain is of course not that repair takes 1.5 seconds to finish.

For example:

The "nodetool repair -pr system_auth" command takes 22 minutes (instead of 3 seconds) #15159 

We have seen in the larger cluster in the filed which took hours to finish the system_auth table repair, e.g., 

```
│ system_auth                   │                role_attributes │ 100%     │ 1h50m53s │
│ system_auth                   │                   role_members │ 100%     │ 1h48m10s │
│ system_auth                   │               role_permissions │ 100%     │ 1h51m13s │
│ system_auth                   │                          roles │ 100%     │ 1h48m23s │
```

Another example, with 5ms delay in dtest:

The time to repair is reduced  from 333s to  0.2s. 

333.26758 s /  0.22625381s = 1472.98

With more nodes and higher latency it will be more visible.