@asias 2023-11-08
> I was impressed by the reduction in RPC calls. rpc_call_nr=115289 (vs 80 after)

Yes, consider a setup with 10ms delay, that would translate to repair time reduction from 

115289 * 10ms = 1152 s

to

80 * 10ms = 0.8s

.