@scylladb-promoter 2023-11-08
#### :red_circle: CI State: FAILURE

:white_check_mark: - Build
:white_check_mark: - Unit Tests
:x: - Sanity Tests
#### Failed Tests (1/21314):
- [test_increment_decrement_counters_in_threads_nodes_restarted](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/4591/testReport/junit/update_cluster_layout_tests/TestUpdateClusterLayout/Tests___Sanity_Tests___test_increment_decrement_counters_in_threads_nodes_restarted) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_increment_decrement_counters_in_threads_nodes_restarted)
#### Build Details:
- Build URL: https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/4591/
- Duration: 2 hr 51 min
- Builder: spider8.cloudius-systems.com