@scylladb-promoter 2023-11-08
#### :green_circle: CI State: SUCCESS

:white_check_mark: - Build
:white_check_mark: - Sanity Tests
:white_check_mark: - Unit Tests
#### Build Details:
- Build URL: https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/4596/
- Duration: 2 hr 5 min
- Builder: spider8.cloudius-systems.com